# Add disable_minification option to CSP config to bypass any post-processing of config data