This PR disables any post-processing of CSP configs when generating a policy. This was prompted by a review of the performance of a typical request. While there are likely plenty of code optimizations that can be made, turning off functionality also makes things go vroom. 

The minification has been buggy and has led to confusion but for those who have a policy that doesn't demand these features, it's wasteful. 

I'll follow up with actual data on how much time is spent in these methods and the impact this could make, but this was the result from a (mostly) empty endpoint. 2.5% of a request cycle to generate a string.

```
        32   (0.7%)          32   (0.7%)     SecureHeaders::ContentSecurityPolicy#dedup_source_list
        27   (0.6%)          27   (0.6%)     SecureHeaders::ContentSecurityPolicy#strip_source_schemes
        29   (0.6%)          26   (0.6%)     SecureHeaders::DynamicConfig#directive_value
        26   (0.6%)          26   (0.6%)     SecureHeaders::DynamicConfig#write_attribute
```

## All PRs:

* [ ] Has tests
* [ ] Documentation updated

## Adding a new header

Generally, adding a new header is always OK.

* Is the header supported by any user agent? If so, which?
* What does it do?
* What are the valid values for the header?
* Where does the specification live?

## Adding a new CSP directive

* Is the directive supported by any user agent? If so, which?
* What does it do?
* What are the valid values for the directive?