@oreoshake 2021-01-22
I snuck a ton of micro optimizations into this PR. I'll separate those out before actually trying to ship this.