@oreoshake 2021-02-04
Just posting some data:

Using this test case for generating a policy (simulating most of what is done in the middleware)

<details>

```ruby
lib = File.expand_path("../lib", __FILE__)
$LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
require "secure_headers"

require 'stackprof'
require 'benchmark/ips'
require 'rack'

request = Rack::Request.new("HTTP_X_FORWARDED_SSL" => "on")
::SecureHeaders::Configuration.default do |config|
  # config.csp.merge!(disable_minification: true)
end

# ::SecureHeaders::Configuration.override(:disable_minification)

puts SecureHeaders.header_hash_for(request)

result = Benchmark.ips do |x|
  x.report "/status" do
    SecureHeaders.header_hash_for(request)
  end
end

output = "stackprof-status.dump"
puts output

StackProf.run(mode: :wall, out: output, raw: true) do
  result.entries[0].iterations.times do
    SecureHeaders.header_hash_for(request)
  end
end
system "stackprof #{output}"

```

</details>

| branch | speed | 
| --- | --- |
| Current default branch | 14.338k (± 1.4%) i/s -     71.750k in   5.005395s |
| This PR | 16.794k (± 1.5%) i/s -     84.350k in   5.023599s |
| This PR with `disable_minification: true` | 20.487k (± 1.6%) i/s -    104.091k in   5.082151s |

As for making modifications to a policy at runtime:

<details>

```ruby
lib = File.expand_path("../lib", __FILE__)
$LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
require "secure_headers"

require 'stackprof'
require 'benchmark/ips'
require 'rack'

request = Rack::Request.new("HTTP_X_FORWARDED_SSL" => "on")
::SecureHeaders::Configuration.default do |config|
  # config.csp.merge!(disable_minification: true)
end

result = Benchmark.ips do |x|
  x.report "/status" do
    SecureHeaders.append_content_security_policy_directives(request, script_src: ["*"])
  end
end

output = "stackprof-status.dump"
puts output

StackProf.run(mode: :wall, out: output, raw: true) do
  result.entries[0].iterations.times do
    SecureHeaders.append_content_security_policy_directives(request, script_src: ["*"])
  end
end
system "stackprof #{output}"

```

</details>

| branch | speed | 
| --- | --- |
| Current default branch | 77.836k (± 2.2%) i/s -    392.598k in   5.046441s |
| This PR | 97.910k (± 2.0%) i/s -    495.414k in   5.061957s |

fyi @ptoomey3 @vcsjones